Moved booking payments logic to separate service
